+++ src/main/java/com/run_us/server/domains/user/model/Penalty.java
@@ -1,4 +1,4 @@
-package com.run_us.server.domain.user.model;
+package com.run_us.server.domains.user.model;
 
 import static com.run_us.server.global.common.GlobalConsts.TIME_ZONE_ID;
