+++ src/main/java/com/run_us/server/domains/user/model/SocialProvider.java
@@ -1,4 +1,4 @@
-package com.run_us.server.domain.user.model;
+package com.run_us.server.domains.user.model;
 
 public enum SocialProvider {
   KAKAO
