+++ src/main/java/com/run_us/server/domains/user/repository/OAuthInfoRepository.java
@@ -1,6 +1,6 @@
-package com.run_us.server.domain.user.repository;
+package com.run_us.server.domains.user.repository;
 
-import com.run_us.server.domain.user.model.OAuthInfo;
+import com.run_us.server.domains.user.model.OAuthInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
