+++ src/main/java/com/run_us/server/domains/user/repository/PenaltyRepository.java
@@ -1,6 +1,6 @@
-package com.run_us.server.domain.user.repository;
+package com.run_us.server.domains.user.repository;
 
-import com.run_us.server.domain.user.model.Penalty;
+import com.run_us.server.domains.user.model.Penalty;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
