+++ src/main/java/com/run_us/server/domains/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.run_us.server.domain.user.repository;
+package com.run_us.server.domains.user.repository;
 
-import com.run_us.server.domain.user.model.User;
+import com.run_us.server.domains.user.model.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
