+++ src/test/java/com/run_us/server/domains/running/RunningFixtures.java
@@ -1,8 +1,8 @@
-package com.run_us.server.domain.running;
+package com.run_us.server.domains.running;
 
-import com.run_us.server.domain.running.model.Running;
-import com.run_us.server.domain.running.model.RunningConstraints;
-import com.run_us.server.domain.running.model.RunningDescription;
+import com.run_us.server.domains.running.domain.Running;
+import com.run_us.server.domains.running.domain.RunningConstraints;
+import com.run_us.server.domains.running.domain.RunningDescription;
 import com.run_us.server.global.utils.PointGenerator;
 import org.locationtech.jts.geom.Point;
