+++ src/test/java/com/run_us/server/domains/user/model/PenaltyRepositoryTest.java
@@ -1,12 +1,11 @@
-package com.run_us.server.domain.user.model;
+package com.run_us.server.domains.user.model;
 
 import static com.run_us.server.global.common.GlobalConsts.TIME_ZONE_ID;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 
-import com.run_us.server.domain.user.repository.PenaltyRepository;
-import com.run_us.server.domain.user.repository.UserRepository;
-import java.time.LocalDate;
+import com.run_us.server.domains.user.repository.PenaltyRepository;
+import com.run_us.server.domains.user.repository.UserRepository;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
