+++ src/test/java/com/run_us/server/domains/user/model/UserFixtures.java
@@ -1,4 +1,4 @@
-package com.run_us.server.domain.user.model;
+package com.run_us.server.domains.user.model;
 
 import java.time.LocalDate;
